refactor: home screen components for gallery ui
