+++ src/renderer/features/home/components/BookmarkTab.tsx
@@ -1,91 +1,96 @@
-import { useAppSelector, useAppDispatch } from "@store/hooks";
 import { useMemo } from "react";
-import BookmarkHistoryListItem from "./BookmarkHistoryListItem";
-import { formatUtils } from "@utils/file";
 import { BookBookmark, LibraryItemWithProgress, MangaBookmark } from "@common/types/db";
-import ListNavigator from "../../components/ListNavigator";
+import { useAppSelector, useAppDispatch } from "@store/hooks";
+import { formatUtils } from "@utils/file";
+import ListNavigator from "@renderer/components/ListNavigator";
+import { setAppSettings } from "@store/appSettings";
 import { faSort } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { setAppSettings } from "@store/appSettings";
 import { useAppContext } from "src/renderer/App";
+import BookmarkHistoryListItem from "./BookmarkHistoryListItem";
 
-const HistoryTab: React.FC = () => {
+const BookmarkTab: React.FC = () => {
+    const bookmarks = useAppSelector((store) => store.bookmarks);
     const library = useAppSelector((store) => store.library);
     const appSettings = useAppSelector((store) => store.appSettings);
     const dispatch = useAppDispatch();
     const { setContextMenuData } = useAppContext();
 
-    const historyItems = useMemo(() => {
-        const items = Object.values(library.items).filter(
-            (item) => item && item.progress,
-        ) as LibraryItemWithProgress[];
-
-        let sorted = [...items];
-
-        if (appSettings.historyListSortBy === "name") {
-            sorted = sorted.sort((a, b) => window.app.betterSortOrder(a.title, b.title));
-        } else {
-            sorted = sorted.sort((a, b) => {
-                const aDate =
-                    a.type === "book"
-                        ? new Date(a.progress?.lastReadAt || 0).getTime()
-                        : new Date(a.progress?.lastReadAt || 0).getTime();
-
-                const bDate =
-                    b.type === "book"
-                        ? new Date(b.progress?.lastReadAt || 0).getTime()
-                        : new Date(b.progress?.lastReadAt || 0).getTime();
-
-                return bDate - aDate;
+    const bookmarksArray = useMemo(() => {
+        const arr: (BookBookmark | MangaBookmark)[] = [];
+        Object.entries(bookmarks.book).forEach(([itemLink, bookmarks]) => {
+            bookmarks?.forEach((bookmark) => {
+                arr.push({ ...bookmark, itemLink });
             });
+        });
+        Object.entries(bookmarks.manga).forEach(([itemLink, bookmarks]) => {
+            bookmarks?.forEach((bookmark) => {
+                arr.push({ ...bookmark, itemLink });
+            });
+        });
+
+        let sorted = [...arr];
+
+        if (appSettings.bookListSortBy === "name") {
+            sorted = sorted.sort((a, b) => {
+                const itemA = library.items[a.itemLink];
+                const itemB = library.items[b.itemLink];
+                if (!itemA || !itemB) return 0;
+                return window.app.betterSortOrder(itemA.title, itemB.title);
+            });
+        } else {
+            sorted = sorted.sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime());
         }
 
-        return appSettings.historyListSortType === "inverse" ? [...sorted].reverse() : sorted;
-    }, [library.items, appSettings.historyListSortBy, appSettings.historyListSortType]);
+        return appSettings.bookListSortType === "inverse" ? [...sorted].reverse() : sorted;
+    }, [bookmarks, library.items, appSettings.bookListSortBy, appSettings.bookListSortType]);
 
-    const filterHistoryItem = (filter: string, item: LibraryItemWithProgress | BookBookmark | MangaBookmark) => {
-        if (!("type" in item)) return false;
+    const filterBookmark = (filter: string, bookmark: LibraryItemWithProgress | BookBookmark | MangaBookmark) => {
+        if ("type" in bookmark) return false;
+        const item = library.items[bookmark.itemLink];
+        if (!item) return false;
+
+        //"page" in bookmark means it's a manga bookmark not epub
+
         const searchText =
-            item.type === "manga"
-                ? item.title +
-                  (formatUtils.files.test(item.progress?.chapterName || "")
-                      ? `${window.path.extname(item.progress?.chapterName || "")}`
-                      : "") +
-                  "manga|manhua|manhwa|webtoon|webcomic|comic"
-                : item.title + ".epub";
+            item.title +
+            (item.progress?.chapterName ? ` ${item.progress.chapterName}` : "") +
+            ("page" in bookmark ? "manga|manhua|manhwa|webtoon|webcomic|comic" : "") +
+            ` ${formatUtils.files.getExt("page" in bookmark ? bookmark.link : bookmark.itemLink)}`;
 
         return new RegExp(filter, "ig").test(searchText);
     };
 
-    const renderHistoryItem = (
-        item: LibraryItemWithProgress | BookBookmark | MangaBookmark,
+    const renderBookmarkItem = (
+        bookmark: LibraryItemWithProgress | BookBookmark | MangaBookmark,
         index: number,
         isSelected: boolean,
     ) =>
-        "type" in item && (
+        "chapterName" in bookmark && (
             <BookmarkHistoryListItem
-                isHistory={true}
-                isBookmark={false}
+                isHistory={false}
+                isBookmark={true}
                 focused={isSelected}
-                link={item.link}
-                id={index}
-                key={`${item.updatedAt}-${index}`}
+                link={bookmark.itemLink}
+                id={bookmark.id}
+                bookmark={bookmark}
+                key={`${bookmark.id}-${bookmark.itemLink}`}
             />
         );
 
-    if (!appSettings.showTabs.history) {
+    if (!appSettings.showTabs.bookmark) {
         return null;
     }
 
     return (
-        <div className="contTab listCont" id="historyTab">
-            <h2>Continue Reading</h2>
+        <div className="contTab listCont" id="bookmarksTab">
+            <h2>Bookmarks</h2>
 
             <ListNavigator.Provider
-                items={historyItems}
-                filterFn={filterHistoryItem}
-                renderItem={renderHistoryItem}
-                emptyMessage="Library Empty"
+                items={bookmarksArray as (BookBookmark | MangaBookmark)[]}
+                filterFn={filterBookmark}
+                renderItem={renderBookmarkItem}
+                emptyMessage="No Bookmarks"
                 onContextMenu={(elem) => elem.dispatchEvent(window.contextMenu.fakeEvent(elem))}
                 onSelect={(elem) => elem.click()}
             >
@@ -95,8 +100,8 @@ const HistoryTab: React.FC = () => {
                             <button
                                 data-tooltip={
                                     "Sort: " +
-                                    (appSettings.historyListSortType === "normal" ? "▲ " : "▼ ") +
-                                    appSettings.historyListSortBy.toUpperCase()
+                                    (appSettings.bookListSortType === "normal" ? "▲ " : "▼ ") +
+                                    appSettings.bookListSortBy.toUpperCase()
                                 }
                                 onClick={(e) => {
                                     const items: Menu.ListItem[] = [
@@ -105,22 +110,22 @@ const HistoryTab: React.FC = () => {
                                             action() {
                                                 dispatch(
                                                     setAppSettings({
-                                                        historyListSortBy: "name",
+                                                        bookListSortBy: "name",
                                                     }),
                                                 );
                                             },
-                                            selected: appSettings.historyListSortBy === "name",
+                                            selected: appSettings.bookListSortBy === "name",
                                         },
                                         {
-                                            label: "Date Updated",
+                                            label: "Date Added",
                                             action() {
                                                 dispatch(
                                                     setAppSettings({
-                                                        historyListSortBy: "date",
+                                                        bookListSortBy: "date",
                                                     }),
                                                 );
                                             },
-                                            selected: appSettings.historyListSortBy === "date",
+                                            selected: appSettings.bookListSortBy === "date",
                                         },
                                         window.contextMenu.template.divider(),
                                         {
@@ -128,22 +133,22 @@ const HistoryTab: React.FC = () => {
                                             action() {
                                                 dispatch(
                                                     setAppSettings({
-                                                        historyListSortType: "normal",
+                                                        bookListSortType: "normal",
                                                     }),
                                                 );
                                             },
-                                            selected: appSettings.historyListSortType === "normal",
+                                            selected: appSettings.bookListSortType === "normal",
                                         },
                                         {
                                             label: "Descending",
                                             action() {
                                                 dispatch(
                                                     setAppSettings({
-                                                        historyListSortType: "inverse",
+                                                        bookListSortType: "inverse",
                                                     }),
                                                 );
                                             },
-                                            selected: appSettings.historyListSortType === "inverse",
+                                            selected: appSettings.bookListSortType === "inverse",
                                         },
                                     ];
                                     setContextMenuData({
@@ -169,4 +174,4 @@ const HistoryTab: React.FC = () => {
     );
 };
 
-export default HistoryTab;
+export default BookmarkTab;
